+++ src/ts/Ads/AdUnits/AdUnitParametersFactory.spec.ts
@@ -47,6 +47,7 @@ describe('AdUnitParametersFactoryTest', () => {
             placement = new Placement();
             thirdPartyEventManagerFactory = ads.ThirdPartyEventManagerFactory;
             adUnitParametersFactory = new AdMobAdUnitParametersFactory(core, ads);
+            placement.getAdUnitId.mockReturnValue('test_ad_unit_1');
         });
 
         describe('when getBaseParameters', () => {
@@ -60,7 +61,9 @@ describe('AdUnitParametersFactoryTest', () => {
                     [ThirdPartyEventMacro.OM_ENABLED]: 'false',
                     [ThirdPartyEventMacro.OM_VENDORS]: '',
                     [ThirdPartyEventMacro.OMIDPARTNER]: 'Unity3d/1.2.10',
-                    [ThirdPartyEventMacro.CACHEBUSTING]: '-1'
+                    [ThirdPartyEventMacro.CACHEBUSTING]: '-1',
+                    [ThirdPartyEventMacro.AD_UNIT_ID_IMPRESSION]: 'test_ad_unit_1',
+                    [ThirdPartyEventMacro.AD_UNIT_ID_OPERATIVE]: 'test_ad_unit_1'
                 });
             });
         });